Namespaces the grsecurity_deb_package var
The build and install roles both include a variable called
`grsecurity_deb_package`, but they mean different things and are handled
differently, so clobbering the values can lead to unexpected failures.
Instead let's adhere to the Ansible best practice of prefixing default
role variables with the name of the role, to avoid clobbering.