Potential fix for failing wait_for task
So I noticed that when trying to run against some of my EC2 instances
that the wait_for task is defaulting to `localhost` and bombs out since
I don't have a local SSH server listening. I re-ordered the host designator
so that inventory_hostname is tried first. I think this has to do with
everything running under the `local_action` module. I also stripped out the
ansible_ssh_host since that var is long-gone in ansible 2.0+