Fixes reboot logic (again)
The changes brought in via #95 did indeed resolve the `wait_for` task
being skipped, but with the unintended side-effect of optimistically
rebooting hosts, even if they were already running the desired kernel
version. That's no good.

Updated the reboot task with conditional logic to skip if the kernel
matches, and chained the subsequent wait tasks on that result.

Closes #94.