Allow settings kernel-img.conf values from vars
This is an optional variable that allows setting up kernel-img.conf properties.
Some of these values in particular were needed during install - specifically,
`clobber_modules` that was the catalyst for adding this option. If the user
doesnt modify this value it will continue to utilize the default conf that is
dropped upon first install.