Loosens logic for finding linux-image deb package
Rather than expecting the default package stem of "linux-image-*", we'll
simply expect the substring "image" to occur in the package name, and
that'll be enough to infer the desired kernel version. The role will
then support installation of kernel packages that have a custom name,
e.g. "organization-purpose-4.4.x-grsec", rather than the more common
(and usually dist-provided) "linux-image-4.4.x-grsec" pattern.

Closes #2.