Address kern ver str match w/ exp. LOCALVERSION
If a kernel is built using `make deb-pkg` and `LOCALVERSION` is passed,
the current jinja filter will not be able to properly parse that
information out. For example, take the following filename:

linux-image-4.4.115-grsec-digitalocean.201802041022_4.4.115-grsec-digitalocean.201802041022-1_amd64.deb

originally the parser only saw this as `4.4.115-grsec` but with the
current commit, it will now extract `4.4.115-grsec-digitalocean.201802041022`. This
matches up with what is set in the kernel uname output and in grub.conf